Merge SubscriptionMonthGiftCard with GiftCard
No need to confuse life with multiple models for the same thing. The
idea before was to only persist attributes related to the certificate
itself and not the purchase of the gift card in general, but this was
rightly challenged at this point and we now have one model for all
aspects of gift cards.