Don't save gift cards that are not paid for
This opens up an edge case where we have charged a customer but their
gift card isn't avilable. If that happens we will get a crash report and
can investigate. The alternative, having gift cards saved even if
they're not paid for will be much worse and is not really an edge case
since cards can get declined for many reasons.