Simplify 3D Secure flow for signups
This change unites code paths for regular and 3D Secure paths for
signups. Most importantly, it avoids the possibility of diverging code
for the different card types. The idea is that 3D Secure is an interrupt
to the process and after completion we retry the same sign up action.
It also removes the intermediate User records for 3D Secure cards and
only saves User records after we've successfully charged in both cases.

As none of this is unit tested, as much of the original code is kept as
possible while still achieving the above goal. Refactoring is for later.