Use ApiKey record for API authentication
Drops hardcoded list of API keys.

Implementation is naive and will query the DB for every request. This
might become a problem, but currently API requests peak at 26 rps which
should be no problem.

This is also the first time that the DB is used for API traffic, which
will lead to API dynos starting to connect to the DB. Current production
configuration is 1 DB connection per worker at 9 workers. Main
deployment uses 6 connections out of the allowed 20 for current DB tier,
meaning we will exceed connection count during rolling deploys but not
at normal. If this is a problem, this change will have to be followed by
configuration or billing changes.