Use explicit regions with URL slugs to determine locale & currency
Regions are the combination of a locale and a currency. These can be
very specific, as with Sweden (Swedish and SEK), or broader, as with
Europe-German (German and EUR).

- Solves locale being different for the same URL's due to setting locale
  based on browser headers. With this change, all pages get one URL for
  each region.
- Uses the same domain for all regions, so we collect SEO reputation on
  our main domain as opposed to spreading it out across subdomains.
- Allows us to offer prices in Euros to English speakers, not just
  German speakers.

As a bonus, this fixes an issue where customers who have cancelled their
plan tries to sign up on a version of the site with another currency
than they originally signed up with. This is not allowed by Stripe and
would give an error previously. Now always shows the original currency
regardless of which region site the user is using.