Recommend regions based on IP & move region switcher to footer
Introduces a notice strip at the top of the screen if the user seems to
be in another region than they're visiting, based on the IPCountry
header from Cloudflare. Dismissable which gets remembered through a
session cookie.

Also moves the region selector to the footer now that we're likely to
recommend the correct region anyway. Switching should become a corner
case. As a bonus, this selector is present even when logged in, which
wasn't the case before.