Reorder homepage with impact numbers on top
Current reasoning is that we want to create a sense that we're doing
this together, so moving numbers of number of users and reduced tonnes
to the top should make that better. This also puts the video at the
bottom since (limited but existign) user testing pointed towards
starting a video being too big of a commitment right when landing on the
page.