Add map with all projects!
Many options for transforming coordinates were considered, but in the
end most had practical problems in terms of how to install dependencies
& the size of required dependencies. Even though including the proj4js
library client side might seem a little excessive for this use, other
options of doing it server-side were prohibitely complicated. A
trade-off indeed. At least we're not including a full scale web mapping
solution on the client :).