Introduce Stimulus & cleanup scroll behavior for home & registrations
We've suffered from having no coherent structure for front-end code for
a while. This introduces Stimulus which gives us a clean way to do it.
Hooks declared right in the HTML make it clear how things behave just
from looking in the template, no additional <script> tags are necessary
for initialization, and controller code can be written as independent
classes.