Rewrite registration to solve numerous issues
I started out with the goal of refactoring this step by step, but since
the difference between the current state and what I was aiming for was
so big, and the current code had so many issues anyway, I decided to go
for a full rewrite.

Known issues solved by this rewrite:

- Errors occurring when attempting to charge could lead to duplicate
Stripe Customers
- Numerous error cases led to incomprehensible error messages behing
shown to users.
- We've also had other support cases where people didn't seem to have
been signed in after starting a subscription.
- Lots of smaller UI issues.