Make gift cards linkable with URL's using keys, stop using session
The session based approach sometimes fail and we end up failing to show
the thank you page for purchases. This uses URL's with keys to avoid
those scenarios, plus makes the code slightly simpler to reason about
and test.

Redirects are updated to take care of the previous scheme. The one
scenario not taken into account is old receipt links, but reloading the
thank you page fixes that and the chance of someone hitting that is
anyway extremely low.