Add new alternative version of sign up page for 25% of visitors
Changes are in line with our hypotethis that we will improve by being
clearer in what you're signing up for as well as showing your calculator
result before having to sign up.

Experiment frequency is 0.5, but since you need to have also being shown
the new calculator for this to show, result is that 25% of users get it.

Co-authored-by: Stefan Lindbohm <stefan@goclimateneutral.org>