Make project optional for invoices
It's important that we register all offset-invoices in our system to keep track of CO2e left total and left in every project. So far the check that everything is registered has been done manually now and then. By adding all invoices to Invoice, even ones without a project, we can easily track which invoices are not registered yet. This will also have the side effect that we can keep track of business income in our admin.