Move community to top of home page to improve relatability
We're pretty clear ourselves on the importance of creating a feeling
that we're doing this together. This was confirmed again in user testing
I did yesterday where the person said "oh, this is very important" to
this section that we actually put way at the bottom of the home page.