Generalize and unify handling of currencies
Switches to lists of prices for each currency, one for consumer prices
and one for business prices, making it simpler to reason about amounts
being charged and to add new currencies.

Also unifies logic for available price steps for subscriptions.
Previously increments of 10 cents or 5 kronor when signing up, but 1
dollar/euro and 5 kronor when changing subscriptions. Now always 50
cents or 5 kronor.

Co-authored-by: Emma Bäckström <emmabckstrm@users.noreply.github.com>