Remove broken "Total received" feature from admin dashboard
This data model is now redundant since we can sum SubscriptionMonths,
GiftCards and FlightOffsets to reach the same result, but actually
better since that will give us the actual sales date instead of the date
when money was paid to us from Stripe. Also good because removing this
means one less place to sort out VAT/non-VAT, which is my current goal.